+++ src/test/java/nextstep/subway/acceptance/AuthSteps.java
@@ -1,10 +1,16 @@
 package nextstep.subway.acceptance;
 
 import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
 import io.restassured.specification.RequestSpecification;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import java.util.HashMap;
+import java.util.Map;
 
 import static nextstep.DataLoader.*;
-import static nextstep.subway.acceptance.MemberSteps.로그인_되어_있음;
 
 public class AuthSteps {
     public static RequestSpecification givenAdminRole() {
@@ -19,5 +25,23 @@ public class AuthSteps {
 
         return RestAssured.given().log().all()
                 .auth().oauth2(userToken);
+    }
+
+    public static String 로그인_되어_있음(String email, String password) {
+        ExtractableResponse<Response> response = 로그인_요청(email, password);
+        return response.jsonPath().getString("accessToken");
+    }
+
+    private static ExtractableResponse<Response> 로그인_요청(String email, String password) {
+        Map<String, String> params = new HashMap<>();
+        params.put("email", email);
+        params.put("password", password);
+
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(params)
+                .when().post("/login/token")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value()).extract();
     }
 }
